- cleanup of logging all loggers under zen.* - conversion of all snmp plugins (except sysedge) - totally unified ssh/telnet/snmp model collection - fixes and updates to the model to reflect more complex collectors
git-svn-id: http://dev.zenoss.com/svnint/trunk/core@14163 f5df01ce-0320-0410-9ed9-ddcd13c732cc

+++ Products/ZenModel/DeviceHW.py
@@ -24,10 +24,10 @@ class DeviceHW(Hardware):
 
     meta_type = "DeviceHW"
 
-    totalMemory = 0.0
+    totalMemory = 0L
 
     _properties = Hardware._properties + (
-        {'id':'totalMemory', 'type':'float', 'mode':'w'},
+        {'id':'totalMemory', 'type':'long', 'mode':'w'},
     )
 
     _relations = Hardware._relations + (
